Fix #124 - Add signal strength legend to Sky View (#140)
* Refactor SkyView into it's own class
* Change GpsSkyFragment to use XML layouts
* Add C/N0 (and SNR for legacy GpsStatus.Listener) signal strength legend to GpsSkyFragment
