+++ GPSTest/src/main/res/drawable/signal_box.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/signal_box">
+    <item android:drawable="@drawable/signal_gradient" />
+    <item>
+        <shape android:shape="rectangle" android:padding="10dp">
+            <corners
+                android:bottomRightRadius="5dp"
+                android:bottomLeftRadius="5dp"
+                android:topLeftRadius="5dp"
+                android:topRightRadius="5dp"/>
+        </shape>
+    </item>
+</layer-list>
