+++ GPSTest/src/main/res/drawable/signal_gradient.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="150dp"
+    android:height="20dp"
+    android:viewportHeight="100"
+    android:viewportWidth="100">
+
+    <path
+        android:pathData="M1,1 H99 V99 H1Z">
+        <aapt:attr name="android:fillColor">
+            <gradient
+                android:endX="99"
+                android:endY="50"
+                android:startX="1"
+                android:startY="50"
+                android:type="linear">
+                <item
+                    android:color="@color/gray"
+                    android:offset="0.0" />
+                <item
+                    android:color="@color/red"
+                    android:offset="0.19" />
+                <item
+                    android:color="@color/yellow"
+                    android:offset="0.66" />
+                <item
+                    android:color="@color/green"
+                    android:offset="1.0" />
+            </gradient>
+        </aapt:attr>
+    </path>
+</vector>
