+++ GPSTest/src/main/res/layout/gps_sky_signal_legend.xml
@@ -1,0 +1,105 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+/*
+** Copyright 2018, Sean J. Barbeau
+**
+** Licensed under the Apache License, Version 2.0 (the "License");
+** you may not use this file except in compliance with the License.
+** You may obtain a copy of the License at
+**
+**     http://www.apache.org/licenses/LICENSE-2.0
+**
+** Unless required by applicable law or agreed to in writing, software
+** distributed under the License is distributed on an "AS IS" BASIS,
+** WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+** See the License for the specific language governing permissions and
+** limitations under the License.
+*/
+-->
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/sky_signal_legend"
+    android:layout_height="wrap_content"
+    android:layout_width="wrap_content">
+
+    <!-- Colored C/N0 legend -->
+    <android.support.v7.widget.AppCompatImageView
+        android:id="@+id/sky_legend_cn0"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerHorizontal="true"
+        android:layout_marginBottom="5dp"
+        android:layout_marginLeft="5dp"
+        android:layout_marginRight="5dp"
+        android:layout_marginTop="5dp"
+        android:background="@android:color/black"
+        android:paddingLeft="1dp"
+        android:paddingRight="1dp"
+        android:paddingTop="2dp"
+        android:paddingBottom="2dp"
+        app:srcCompat="@drawable/signal_box" />
+    <!-- CN/0 legend scale text and lines -->
+    <View
+        android:id="@+id/sky_legend_cn0_left_line"
+        android:layout_width="1dp"
+        android:layout_height="10dp"
+        android:layout_below="@id/sky_legend_cn0"
+        android:layout_alignLeft="@id/sky_legend_cn0"
+        android:layout_alignStart="@id/sky_legend_cn0"
+        android:layout_margin="2dp"
+        android:background="@android:color/black" />
+
+    <TextView
+        android:id="@+id/sky_legend_cn0_left_text"
+        style="@style/sky_legend_value"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/sky_legend_cn0_left_line"
+        android:layout_gravity="start"
+        android:layout_alignLeft="@+id/sky_legend_cn0"
+        android:layout_alignStart="@+id/sky_legend_cn0"
+        android:text="@string/sky_legend_cn0_low"
+        android:textAlignment="textStart" />
+
+    <View
+        android:id="@+id/sky_legend_cn0_center_line"
+        android:layout_width="1dp"
+        android:layout_height="10dp"
+        android:layout_below="@id/sky_legend_cn0"
+        android:layout_centerHorizontal="true"
+        android:layout_margin="2dp"
+        android:background="@android:color/black" />
+
+    <TextView
+        android:id="@+id/sky_legend_cn0_center_text"
+        style="@style/sky_legend_value"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/sky_legend_cn0_center_line"
+        android:layout_centerHorizontal="true"
+        android:layout_gravity="center"
+        android:text="@string/sky_legend_cn0_middle"
+        android:textAlignment="center" />
+
+    <View
+        android:id="@+id/sky_legend_cn0_right_line"
+        android:layout_width="1dp"
+        android:layout_height="10dp"
+        android:layout_below="@id/sky_legend_cn0"
+        android:layout_alignRight="@id/sky_legend_cn0"
+        android:layout_alignEnd="@id/sky_legend_cn0"
+        android:layout_margin="2dp"
+        android:background="@android:color/black" />
+
+    <TextView
+        android:id="@+id/sky_legend_cn0_right_text"
+        style="@style/sky_legend_value"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/sky_legend_cn0_left_line"
+        android:layout_gravity="start"
+        android:layout_alignRight="@+id/sky_legend_cn0"
+        android:layout_alignEnd="@+id/sky_legend_cn0"
+        android:paddingRight="1dp"
+        android:text="@string/sky_legend_cn0_high"
+        android:textAlignment="textStart" />
+</RelativeLayout>
