+++ GPSTest/src/main/res/layout/gps_sky_signal_legend_title.xml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+/*
+** Copyright 2018, Sean J. Barbeau
+**
+** Licensed under the Apache License, Version 2.0 (the "License");
+** you may not use this file except in compliance with the License.
+** You may obtain a copy of the License at
+**
+**     http://www.apache.org/licenses/LICENSE-2.0
+**
+** Unless required by applicable law or agreed to in writing, software
+** distributed under the License is distributed on an "AS IS" BASIS,
+** WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+** See the License for the specific language governing permissions and
+** limitations under the License.
+*/
+-->
+<!-- C/N0 legend title and units-->
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/sky_legend_cn0_title_layout"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:orientation="vertical" >
+    <TextView
+        android:id="@+id/sky_legend_cn0_title"
+        style="@style/sky_legend_label"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:text="@string/gps_cn0_column_label"
+        android:textAlignment="center" />
+    <TextView
+        android:id="@+id/sky_legend_cn0_units"
+        style="@style/sky_legend_value"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:text="@string/sky_legend_cn0_units"
+        android:textAlignment="center" />
+</LinearLayout>
