+++ GPSTest/src/main/res/values/colors.xml
@@ -42,4 +42,8 @@
 
     <!-- GNSS Sky View -->
     <color name="not_in_view_sat">#11000000</color>
+    <color name="gray">#ff888888</color>
+    <color name="red">#ffff0000</color>
+    <color name="yellow">#ffffff00</color>
+    <color name="green">#ff00ff00</color>
 </resources>
